requirements.txt
# Core LLM and Framework
openai>=1.0.0
llama-index>=0.9.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
llama-index-vector-stores-faiss>=0.1.0

# Vector Databases and Embeddings
faiss-cpu>=1.7.4
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
InstructorEmbedding>=1.0.1
voyageai>=0.2.0

# Code Parsing
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4

# Search and Retrieval
rank-bm25>=0.2.2
nltk>=3.8.1

# Web Framework for Documentation Interface
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Code Analysis
gitpython>=3.1.40
pygments>=2.17.0
black>=23.12.0
pylint>=3.0.0

# Evaluation and Metrics
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
toml>=0.10.2
rich>=13.7.0

# Optional: Visualization for dependency graphs
networkx>=3.2.0
matplotlib>=3.8.0
graphviz>=0.20.1
